Consume TravelMission into VehicleMission
Changes:
- VehicleMission subclassed directly from Mission
- Travel logic moved into VehicleMission
- Deleted TravelMission
- Updated callers of TravelMission

mars-sim#547
