+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/tool/map/NavpointMapLayer.java
@@ -15,7 +15,7 @@ import org.mars_sim.msp.core.Coordinates;
 import org.mars_sim.msp.core.IntPoint;
 import org.mars_sim.msp.core.person.ai.mission.Mission;
 import org.mars_sim.msp.core.person.ai.mission.NavPoint;
-import org.mars_sim.msp.core.person.ai.mission.TravelMission;
+import org.mars_sim.msp.core.person.ai.mission.VehicleMission;
 import org.mars_sim.msp.ui.swing.ImageLoader;
 
 /**
@@ -85,13 +85,13 @@ public class NavpointMapLayer implements MapLayer {
 	 */
 	public void displayLayer(Coordinates mapCenter, String mapType, Graphics g) {
 		if (singleMission != null) {
-			if (singleMission instanceof TravelMission)
-				displayMission((TravelMission) singleMission, mapCenter, mapType, g);
+			if (singleMission instanceof VehicleMission)
+				displayMission((VehicleMission) singleMission, mapCenter, mapType, g);
 		} else {
 			// MissionManager manager = Simulation.instance().getMissionManager();
 			for (Mission mission : missionManager.getMissions()) {
-				if (mission instanceof TravelMission)
-					displayMission((TravelMission) mission, mapCenter, mapType, g);
+				if (mission instanceof VehicleMission)
+					displayMission((VehicleMission) mission, mapCenter, mapType, g);
 			}
 		}
 
@@ -108,7 +108,7 @@ public class NavpointMapLayer implements MapLayer {
 	 * @param mapType   the type of map.
 	 * @param g         graphics context of the map display.
 	 */
-	private void displayMission(TravelMission mission, Coordinates mapCenter, String mapType, Graphics g) {
+	private void displayMission(VehicleMission mission, Coordinates mapCenter, String mapType, Graphics g) {
 		for (int x = 0; x < mission.getNumberOfNavpoints(); x++) {
 			// NavPoint navpoint = mission.getNavpoint(x);
 			displayNavpoint(mission.getNavpoint(x), mapCenter, mapType, g);
